pyproject.toml
[project]
name = "streamlit-chatbot-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-core>=0.3.45",
    "langchain-openai>=0.3.8",
    "numpy>=2.2.3",
    "openai>=1.66.3",
    "pdfplumber>=0.11.5",
    "pypdf>=5.3.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "streamlit>=1.43.2",
    "tiktoken>=0.9.0",
    "transformers>=4.49.0",
]
